+++ R/identityWordcount.R
@@ -17,7 +17,7 @@ identityWordcount <- function(orgn.df, ntop = 50){
   # return only 'ntop' words
   # in worcloud2() use:
   #   size = 0.2, ellipticity = 0.5
-
+  
   # Sort by 'pident', keep highest unique scientific_name match
   orgn.df <- orgn.df[ order(orgn.df$pident, decreasing = T), ]
   orgn.df <- orgn.df[ !duplicated(orgn.df$scientific_name), ]