+++ scripts/palmid.sh
@@ -130,6 +130,20 @@ palmscan -search_pp $INPUT -loconf -rdrp \
   -fevout $OUTDIR/$OUTNAME.fev \
   -ppout $OUTDIR/$OUTNAME.trim.fa
 
+# hack to fix qlen to total input query length
+# (but we now assume aa-sequence input)
+seqkit stats -T $INPUT | tail -n-1 - \
+  > fa.stats
+SEQTYPE=$(cut -f 3 fa.stats)
+SEQLEN=$(cut -f 5  fa.stats)
+
+if [ "$SEQTYPE" == "DNA" ]
+then
+  SEQLEN=$(($SEQLEN*3))
+fi
+
+sed -i "s/qlen=[0-9]*/qlen=$SEQLEN/g" $OUTDIR/$OUTNAME.fev
+
 echo ''
 echo ' palmprint:'
 echo ''
@@ -213,8 +227,7 @@ HTML_OUTPUT="'$OUTDIR/$OUTNAME.html'"
 
 cp /home/palmid/palmid.Rmd /tmp/palmid.Rmd
 mkdir -p /tmp/img
-cp /home/palmid/data/tax_legend.png /tmp/img/tax_legend.png
-
+cp /home/palmid/data/*.png /tmp/img/
 
 Rscript -e "rmarkdown::render( \
   input = '/tmp/palmid.Rmd', \